test when args>0

+++ src/main/java/com/project/test/UserTest.java
@@ -17,35 +17,36 @@ public class UserTest implements CommandLineRunner{
 	
     @Override
     public void run(String...args) throws Exception {
-    	
-        User u1 = new User();
-        u1.setName("phil");
-        u1.setUsername("philuser");
-        u1.setIs_admin(false);        
-        User u2 = new User();
-        u2.setName("karen");
-        u2.setUsername("karenuser");
-        u2.setIs_admin(true);
-        
-        userService.add(u1);
-        userService.add(u2);
-        
-        List<User> userlist = userService.getUsers();
-        for(User u : userlist)
-        	System.out.println(u);
-        
-        User u3 = userService.getUserById(userlist.get(0).getId());
-        u3.setName("pete");
-        u3.setUsername("peteuser");
-        
-        userService.add(u3);
-        
-        userlist = userService.getUsers();
-        for(User u : userlist)
-        {
-        	System.out.println(u);
-        	userService.delete(u.getId());
-        }
-        
+    	if(args.length > 0)
+    	{
+	        User u1 = new User();
+	        u1.setName("phil");
+	        u1.setUsername("philuser");
+	        u1.setIs_admin(false);        
+	        User u2 = new User();
+	        u2.setName("karen");
+	        u2.setUsername("karenuser");
+	        u2.setIs_admin(true);
+	        
+	        userService.add(u1);
+	        userService.add(u2);
+	        
+	        List<User> userlist = userService.getUsers();
+	        for(User u : userlist)
+	        	System.out.println(u);
+	        
+	        User u3 = userService.getUserById(userlist.get(0).getId());
+	        u3.setName("pete");
+	        u3.setUsername("peteuser");
+	        
+	        userService.add(u3);
+	        
+	        userlist = userService.getUsers();
+	        for(User u : userlist)
+	        {
+	        	System.out.println(u);
+	        	userService.delete(u.getId());
+	        }
+    	}
     }
 }
